pyproject.toml
[project]
name = "mnist-project"
version = "0.1.0"
description = "Interactive web application for handwritten digit recognition using deep learning and Streamlit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "streamlit>=1.51.0",
    "streamlit-drawable-canvas>=0.9.3",
    "tensorflow>=2.20.0",
]
